Refactoring of plugins code
