+++ app/code/Magento/Backend/etc/adminhtml/di.xml
@@ -59,7 +59,7 @@
             </argument>
         </arguments>
     </type>
-    <type name="Magento\Backend\App\ActionInterface">
+    <type name="Magento\Framework\App\ActionInterface">
         <plugin name="admin_authentication_plugin" type="Magento\Backend\App\Action\Plugin\BackendActionAuthenticationPlugin" sortOrder="100"/>
         <plugin name="admin_massaction_key_plugin" type="Magento\Backend\App\Action\Plugin\BackendActionMassactionKeyPlugin" sortOrder="11"/>
         <plugin name="admin_loaddesign_plugin" type="Magento\Backend\App\Action\Plugin\BackendActionLoadDesignPlugin" sortOrder="101"/>
